Split view.h
